requirements.txt
google-cloud-aiplatform
google-cloud-firestore
google-cloud-secret-manager
feedparser
requests
python-dotenv
